+++ structured_derby/src/main/java/com/splicemachine/derby/impl/sql/execute/operations/ScanOperation.java
@@ -254,7 +254,7 @@ public abstract class ScanOperation extends SpliceBaseOperation implements Curso
     }
 
     @Override
-	public int[] getRootAccessedCols(int tableNumber) {
+	public int[] getRootAccessedCols(long tableNumber) {
         return baseColumnMap;
 	}
